Added --checksum option in rsync
Problem :
rsync skips to copy files when the files have identical sizes and modification times on the source
and destination hosts

Solution :
rysnc with -c or --checksum option can be used to copy the files that have identical sizes and
modification times on the source and destination hosts. It modifies the file-times-and-sizes
heuristic, but here it ignores times and examines only sizes. Files on the source and destination
hosts with either a changed size or a changed checksum are selected for transfer.

Implementation :
Added the -c or --checksum option with rsync command everywhere in cm utlity codebase

Added behave tests for the below manual test steps :
1. create a file with "hostname > tmp.txt" on all the hosts using gpssh
2. get the modification time of tmp.txt on source host with "stat -c '%y' tmp.txt",
3. update the modification time of tmp.txt on target host with "touch -m -d "date_time" tmp.txt
4. run gpsync of tmp.txt file on all the hosts without checksum flag
5. make sure that files are not synced
6. run gpsync of tmp.txt file on all the hosts with checksum flag
7. make sure files are synced