gpcheckcat: Add new option -x to set session level GUCs
Problem : Running gpcheckcat FK check taking a lot of time to complete, One of the workarounds to
fix this is to Pull the FK query and run it separately in psql using the GUC disable_cost='1e20'
at session level. It help in making the query run faster but its not acceptable solution.So need to set the
GUC at session level.

Solution : Added a new option [-x] in gpcheckcat that will provide support for PGOPTIONS to pass multiple
session GUCs [such as disable_costs, gp_role] to set the parameter.

Implementation : Added a new option -x which is used to pass the multiple session GUCs to set the parameter.
This is helpful when wanting to run certain tests in gpcheckcat with session level parameters.
User can use the -x option multiple times to specify multiple GUC
we can pass the GUC like : gpcheckcat -x disable_cost=3e15 -x log_min_messages=debug2 -R foreign_key

Added behave test case for the same