Added --ignore-times and --whole-file option in rsync
Fixes https://github.com/greenplum-db/gpdb/issues/14744

Problem :
rsync skips file if it has identical sizes and modification times on the source and destination
hosts, this makes rsync behave differently than scp. gpscp was replaced with gpsync (which uses
rsync to copy files) but we don't want to change the behaviour.

Solution :
To make gpsync behave the same as gpscp, -I or --ignore-times and -W or --whole-file option can
be used with rsync. It will copy the files that have identical size and modification time on the
source and destination hosts. -I or --ignore-times is used to make sure files that match in size
and mod-time are not skipped. -W or --whole-file flag used to copy files completely without using
delta-transfer algorithm.

Implementation :
Added the -I or --ignore-times and -W or --whole-file option with rsync command everywhere in
cluster management utility codebase(wherever scp was used to copy files) to make sure rsync
behaves exactly same as scp.
Note: gpsync utility taken care as a part of separete PR