Fix gpstart issues for large segment host
Issue : While running gpstart with 64 segment nodes with 384 primary and 384 mirror,
a warning message is displayed: "[WARNING]:-Number of segments which failed to start: 384."
However, when checking with "gpstate," all segments appear to be in a normal state.

RCA : When the batch size (-B) is less than the number of segments host gpstart starts
the mirror segment before the primary segment. From segment logs, we found out pg_ctl
-w start for mirror is timeout out with a nonzero exit status. This leads to a warning
in gpstart for mirror segments - [WARNING]:-Number of segments which failed to start: 384.
The PR(https://github.com/greenplum-db/gpdb/pull/15994) introduced the changed pg_ctl start
with -w flag for Mirrors now wait until walreceiver is started to be marked as up - which can
only happen if primary is up due to these mirror segments leading to the timeout issue.
The mirror start operations still runs in the background and when its primary is up mirror
is also started this leads to a warning in gpstart for mirror segments.

Solution - Implement the changes to start the primary before the mirror, resolving the timeout problem.